// src/types/partials/Reference.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TUri } from '../simpleTypes/Uri';
import { TIdentifier } from '../partials/Identifier';

export type TReference = {
  id?: string;
  extension?: TExtension[];
  reference?: string;
  type?: TUri;
  identifier?: TIdentifier;
  display?: string;
  _sourceAssigningAuthority?: string;
  _uuid?: string;
  _sourceId?: string;
};