// src/types/resources/Organization.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TContactPoint } from '../partials/ContactPoint';
import { TAddress } from '../partials/Address';
import { TReference } from '../partials/Reference';
import { TOrganizationContact } from '../partials/OrganizationContact';

export type TOrganization = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  active?: boolean;
  type?: TCodeableConcept[];
  name?: string;
  alias?: string[];
  telecom?: TContactPoint[];
  address?: TAddress[];
  partOf?: TReference;
  contact?: TOrganizationContact[];
  endpoint?: TReference[];
};