// src/types/resources/Medication.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TRatio } from '../partials/Ratio';
import { TMedicationIngredient } from '../partials/MedicationIngredient';
import { TMedicationBatch } from '../partials/MedicationBatch';

export type TMedication = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  code?: TCodeableConcept;
  status?: string;
  manufacturer?: TReference;
  form?: TCodeableConcept;
  amount?: TRatio;
  ingredient?: TMedicationIngredient[];
  batch?: TMedicationBatch;
};