// src/types/partials/Quantity.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TDecimal } from '../simpleTypes/Decimal';
import { TUri } from '../simpleTypes/Uri';

export type TQuantity = {
  id?: string;
  extension?: TExtension[];
  value?: TDecimal;
  comparator?: string;
  unit?: string;
  system?: TUri;
  code?: string;
};